--- crates/mpc-protocols/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "mpc-protocols"
-edition = "2021"
-version.workspace = true
-
-[dependencies]
-serde.workspace = true
-serde_json.workspace = true
-schemars.workspace = true
-pera-types.workspace = true
-move-core-types.workspace = true
-anyhow.workspace = true
-tokio.workspace = true
-bcs.workspace = true
