--- crates/mpc-protocols/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod mpc_events;
-pub mod mpc_service;
