+++ crates/pera-analytics-indexer/Cargo.toml
@@ -51,7 +51,7 @@ pera-json-rpc-types.workspace = true
 pera-package-resolver.workspace = true
 simulacrum.workspace = true
 arrow.workspace = true
-gcp-bigquery-client = "0.18.0"
+gcp-bigquery-client = "0.23.0"
 snowflake-api.workspace = true
 tap.workspace = true
