+++ crates/pera-core/src/lib.rs
@@ -78,4 +78,5 @@ mod type_param_tests;
 pub mod signature_verifier;
 
 pub mod runtime;
+pub mod signature_mpc;
 mod transaction_signing_filter;
